% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dir}
\alias{check_dir}
\title{Check if a Directory Exists and Create if Not}
\usage{
check_dir(dir)
}
\arguments{
\item{dir}{Character string path to a directory.}
}
\description{
Check if a Directory Exists and Create if Not
}
