% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{crowdforecast_data}
\alias{crowdforecast_data}
\title{Crowd forecast data used for the paper}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 51244 rows and 11 columns.
}
\usage{
crowdforecast_data
}
\description{
Crowd forecast data used for the paper
}
\keyword{datasets}
