% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dailytruth_data}
\alias{dailytruth_data}
\title{Daily truth data used for the paper}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2200 rows and 5 columns.
}
\usage{
dailytruth_data
}
\description{
Daily truth data used for the paper
}
\keyword{datasets}
