% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dates_to_epiweek}
\alias{dates_to_epiweek}
\title{Dates to EpiWeeks}
\usage{
dates_to_epiweek(df)
}
\arguments{
\item{df}{An input dataframe with a \code{date} column.}
}
\value{
A dataframe.
}
\description{
Dates to EpiWeeks
}
