% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ensemble_members}
\alias{ensemble_members}
\title{Models included in the official hub ensemble}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 273 rows and 6 columns.
}
\usage{
ensemble_members
}
\description{
Models included in the official hub ensemble
}
\keyword{datasets}
