% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ensemble_models}
\alias{ensemble_models}
\title{Names of all ensemble models}
\format{
An object of class \code{character} of length 12.
}
\usage{
ensemble_models
}
\description{
Names of all ensemble models
}
\keyword{datasets}
