% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{epitrend}
\alias{epitrend}
\title{Classification of the epidemic into falling, rising etc (not used)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 314 rows and 7 columns.
}
\usage{
epitrend
}
\description{
Classification of the epidemic into falling, rising etc (not used)
}
\keyword{datasets}
