% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow2.R
\name{extract_samples}
\alias{extract_samples}
\title{Extract Simulated Samples}
\usage{
extract_samples(output, target)
}
\arguments{
\item{output}{A list as output by \code{simulate_crowd_cases}.}

\item{target}{A character string indicating the target.}
}
\value{
A dataframe of sampled values by date.
}
\description{
Extract Simulated Samples
}
