% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{filtered_data}
\alias{filtered_data}
\title{Pre-filtered data used for the paper (with death forecasts restricted to the period after December 14th 2020)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 119600 rows and 15 columns.
}
\usage{
filtered_data
}
\description{
Pre-filtered data used for the paper (with death forecasts restricted to the period after December 14th 2020)
}
\keyword{datasets}
