% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{forecast_dates}
\alias{forecast_dates}
\title{Forecast dates used for the paper}
\format{
An object of class \code{list} of length 4.
}
\usage{
forecast_dates
}
\description{
Forecast dates used for the paper
}
\keyword{datasets}
