% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations.R
\name{get_observations}
\alias{get_observations}
\title{Get Observed Cases linked with Deaths}
\usage{
get_observations(dir, target_date = Sys.Date(), weeks = 8, locs)
}
\arguments{
\item{dir}{A character string indicating the path to the target data folder.}

\item{target_date}{Date or character string indicating the target forecast
date.}

\item{weeks}{Numeric, defaults to 8. Number of weeks of data to include.}

\item{locs}{An optional character vector to use to filter locations by name.}
}
\value{
A dataframe of primary and secondary observations
}
\description{
Get Observed Cases linked with Deaths
}
