% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{google_auth}
\alias{google_auth}
\title{Authentificate for Google Sheets}
\usage{
google_auth(
  service_account = "default",
  email = "epiforecasts@gmail.com",
  cache_folder = ".secrets"
)
}
\arguments{
\item{service_account}{the path to a JSON file that has all the information
of the Google service account. If a service account is presented it will
be used and the other arguments will be ignored}

\item{cache_folder}{path to the folder where secrets for the email address
provided are stored.}

\item{enail}{alternatively, provide the email address to an authorised
account}
}
\description{
Authentificate for Google Sheets
}
