% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow2.R
\name{load_epinow}
\alias{load_epinow}
\title{Load a EpiNow2 Model}
\usage{
load_epinow(target_region, dir, date)
}
\arguments{
\item{target_region}{A dataframe containing a forecast as produced by
\code{format_forecast}.}

\item{dir}{Character string indicating the location name.}

\item{date}{Character vecetor, indicates target regions.}
}
\value{
A EpiNow2 model
}
\description{
Load a EpiNow2 Model
}
