% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{locations}
\alias{locations}
\title{Location and Population Look Up for Germany and Poland}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 34 rows and 3 columns.
}
\usage{
locations
}
\description{
Downloaded from https://github.com/KITmetricslab/covid19-forecast-hub-de
}
\keyword{datasets}
