% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_cumulative}
\alias{make_cumulative}
\title{Make Data Cumulative}
\usage{
make_cumulative(inc)
}
\arguments{
\item{inc}{A data frame comtaining: \code{target_end_date}, \code{location},
\code{location_name}.}
}
\value{
A cumualtive weekly data frame.
}
\description{
Make Data Cumulative
}
