% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_weekly}
\alias{make_weekly}
\title{Make Incidence data weekly}
\usage{
make_weekly(inc)
}
\arguments{
\item{inc}{A data frame containing: \code{location}, \code{location_name}, and
\code{epiweek}.}
}
\value{
A data frame grouped by week.
}
\description{
Make Incidence data weekly
}
