% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{prediction_data}
\alias{prediction_data}
\title{Forecast data used for the paper}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 186944 rows and 9 columns.
}
\usage{
prediction_data
}
\description{
Forecast data used for the paper
}
\keyword{datasets}
