% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{regular_models}
\alias{regular_models}
\title{Names of all regular models}
\format{
An object of class \code{character} of length 4.
}
\usage{
regular_models
}
\description{
Names of all regular models
}
\keyword{datasets}
