% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_forecast}
\alias{save_forecast}
\title{Save a Forecast}
\usage{
save_forecast(forecast, loc_name, loc, type = "", date, folder, model)
}
\arguments{
\item{forecast}{A dataframe containing a forecast as produced by
\code{format_forecast}.}

\item{loc_name}{Character string indicating the location name.}

\item{loc}{Character vecetor, indicates target regions.}

\item{type}{Character string default to "". Indicates the target type.}

\item{date}{A character string or Date indicating the date of forecast.}

\item{folder}{Character string indicating the target folder.}

\item{model}{Character string indicating the model name.}
}
\description{
Save a Forecast
}
