% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow2.R
\name{simulate_crowd_cases}
\alias{simulate_crowd_cases}
\title{Simulate Cases using Crowd Reproduction Number estimates}
\usage{
simulate_crowd_cases(crowd_rt, model_dir, target_date)
}
\arguments{
\item{crowd_rt}{A data frame containing the following variables:
\code{location}, \code{target}, \code{date}, \code{value}, and \code{sample}.}

\item{model_dir}{A character string giving the path to the directory
in which the saved EpiNow2 model objects are stored.}

\item{target_date}{A character string indicating the target date.}
}
\value{
A list containing the output from EpiNow2::simulate_infections
named by target and location.
}
\description{
Simulate Cases using Crowd Reproduction Number estimates
}
