% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{truth_data}
\alias{truth_data}
\title{Truth data used for the paper}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 314 rows and 4 columns.
}
\usage{
truth_data
}
\description{
Truth data used for the paper
}
\keyword{datasets}
