% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_and_wait}
\alias{try_and_wait}
\title{Attempt to Execute an Expression and Retry After Failure
#'}
\usage{
try_and_wait(expr, time_to_wait = 120, number_of_attempts = 10)
}
\arguments{
\item{expr}{an expression that shell be executed}

\item{time_to_wait}{time to wait until the next try after a failure}

\item{number_of_attempts}{numeric, how often shall we try?}
}
\value{
outcome of the expression to be evaluated
}
\description{
Attempt to Execute an Expression and Retry After Failure
#'
}
