% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{unfiltered_data}
\alias{unfiltered_data}
\title{Unfiltered version of the combined prediction and truth data used for the paper}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 145544 rows and 15 columns.
}
\usage{
unfiltered_data
}
\description{
Unfiltered version of the combined prediction and truth data used for the paper
}
\keyword{datasets}
