% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{locations}
\alias{locations}
\alias{truth_data}
\alias{dailytruth_data}
\alias{prediction_data}
\alias{crowdforecast_data}
\alias{combined_data}
\alias{forecast_dates}
\alias{filtered_data}
\alias{unfiltered_data}
\alias{epitrend}
\title{Location and Population Look Up for Germany and Poland}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 34 rows and 3 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 262 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 1834 rows and 5 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 52440 rows and 9 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 73508 rows and 10 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 52486 rows and 10 columns.

An object of class \code{list} of length 4.

An object of class \code{data.table} (inherits from \code{data.frame}) with 26496 rows and 15 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 33304 rows and 15 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 262 rows and 6 columns.
}
\usage{
locations

truth_data

dailytruth_data

prediction_data

crowdforecast_data

combined_data

forecast_dates

filtered_data

unfiltered_data

epitrend
}
\description{
Downloaded from https://github.com/KITmetricslab/covid19-forecast-hub-de
}
\keyword{datasets}
